+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - This is first version using Bitmovin Player iOS V3 SDK. This is not backward compatible with V2 player SDK.
 
+### Known Issues
+- Playlist API available in V3 player SDK is not supported.
+
 ## [1.4.0]
 
 ### Fixed